refactor: ReservationPageController 책임 분리
